+++ Assets/Scripts/EventBus/EventsPP.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+public struct ConversationEndEvent : IEventPP
+{
+
+}
