--- Assets/Scripts/Interfaces/IEventPP.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 6acf35a65918ef846a07e5ad0dff9234
